Extract init iau sdk to function
